% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissCP.R
\name{BIC}
\alias{BIC}
\title{BIC}
\usage{
BIC(residual, phi)
}
\arguments{
\item{residual}{residual matrix}

\item{phi}{estimated coefficient matrix of the model}
}
\value{
A list object, which contains the followings
\describe{
  \item{BIC}{BIC value}
  \item{HBIC}{HBIC value}
}
}
\description{
BIC  and HBIC function
}
