% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissCP.R
\name{BIC_threshold}
\alias{BIC_threshold}
\title{BIC_threshold}
\usage{
BIC_threshold(
  beta.final,
  k,
  m.hat,
  brk,
  data_y,
  data_x = NULL,
  b_n = 2,
  nlam = 20
)
}
\arguments{
\item{beta.final}{estimated parameter coefficient matrices}

\item{k}{dimensions of parameter coefficient matrices}

\item{m.hat}{number of estimated change points}

\item{brk}{vector of estimated change points}

\item{data_y}{input data matrix (response), with each column representing the time series component}

\item{data_x}{input data matrix (predictor), with each column 1}

\item{b_n}{the block size}

\item{nlam}{number of hyperparameters for grid search}
}
\value{
lambda.val.best, the tuning parameter lambda selected by BIC.
}
\description{
BIC threshold for final parameter estimation
}
