% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissCP.R
\name{constant_generation}
\alias{constant_generation}
\title{constant_generation}
\usage{
constant_generation(n, p, d, vns, brk)
}
\arguments{
\item{n}{the sample size}

\item{p}{the data dimension}

\item{d}{the number of nonzero coeddficients}

\item{vns}{the jump size. It can be a vector or a single value. If single value, it is same for all break points}

\item{brk}{the break points' locations}
}
\value{
the parameter matrix used to generate data
}
\description{
function to generate constant given jump size and break points
}
