% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissCP.R
\name{data_generation}
\alias{data_generation}
\title{data_generation}
\usage{
data_generation(n, mu, sigma, brk = n + 1)
}
\arguments{
\item{n}{the number of data points}

\item{mu}{the matrix of mean parameter}

\item{sigma}{covariance matrix of the white noise}

\item{brk}{vector of change points}
}
\value{
data_y matrix of generated mean shift data
}
\description{
The function to generate mean shift data
}
