% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissCP.R
\name{imputation}
\alias{imputation}
\title{imputation}
\usage{
imputation(data, block.size)
}
\arguments{
\item{data}{data before the imputation}

\item{block.size}{the block size that are used to impute the missing}
}
\value{
the data matrix without missing values after imputation
}
\description{
function to do the imputation based on block size
}
