% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissCP.R
\name{imputation2}
\alias{imputation2}
\title{imputation2}
\usage{
imputation2(data, cp.candidate)
}
\arguments{
\item{data}{data before the imputation}

\item{cp.candidate}{the change point candidate that are used to impute the missing}
}
\value{
the data matrix without missing values after imputation
}
\description{
function to do the imputation based on change point candidate
}
