% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissCP.R
\name{MCAR}
\alias{MCAR}
\title{MCAR}
\usage{
MCAR(data, alpha)
}
\arguments{
\item{data}{data before the missing case}

\item{alpha}{the percentage of missing compared to whole data}
}
\value{
the data matrix with missing values
}
\description{
function to do the missing assuming the missing completely at random
}
