% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MissCP.R
\name{second.step}
\alias{second.step}
\title{second.step}
\usage{
second.step(
  data_y,
  data_x,
  max.iteration = max.iteration,
  tol = tol,
  cp.first,
  beta.est,
  blocks,
  data_y_miss
)
}
\arguments{
\item{data_y}{input data matrix, with each column representing the time series component}

\item{data_x}{input data matrix}

\item{max.iteration}{max number of iteration for the fused lasso}

\item{tol}{tolerance for the fused lasso}

\item{cp.first}{the selected break points after the first step}

\item{beta.est}{the estiamted parameters by block fused lasso}

\item{blocks}{the blocks}

\item{data_y_miss}{the data y matrix before the first imputation}
}
\value{
A list object, which contains the followings
\describe{
  \item{cp.final}{a set of selected break point after the exhaustive search step}
  \item{beta.hat.list}{the estimated coefficient matrix for each segmentation}
}
}
\description{
Reimputate the missing values and perform the exhaustive search to "thin out" redundant break points.
}
